requirements_avancement.txt
discord.py>=2.6.0,<2.8.0
aiohttp>=3.8.1